Driver#connect(): phpdoc fixes
